requirements.txt
Django<1.9
Pillow
amqp
anyjson
billiard
biplist
celery
defusedxml
django-celery
django-debug-toolbar
git+https://github.com/brianz/django-debug-toolbar-mongo.git
django-debug-toolbar-template-profiler
django-debug-toolbar-template-timings
django-debug-toolbar-vcs-info
django-tastypie==0.13.3
git+https://github.com/frbapolkosnik/django-tastypie-mongoengine.git
future
kombu
lxml
mongoengine==0.10.6
git+https://github.com/MongoEngine/django-mongoengine.git
pip>=8.1.2
git+https://github.com/kbandla/pydeep.git
pymongo==3.2.2
pyparsing
python-dateutil
python-ldap
python-magic==0.4.6
python-mimeparse
pytz
pyyaml
requests
setuptools>=28.3.0
simplejson
six
ushlex
wsgiref
olefile
pyimpfuzzy
m2crypto
